refactor: customerDto를 request와 response로 구분지어 패키징
