+++ src/main/java/woowacourse/shoppingcart/dto/customer/request/EmailDuplicateRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto.customer;
+package woowacourse.shoppingcart.dto.customer.request;
 
 import javax.validation.constraints.Pattern;
