+++ src/main/java/woowacourse/shoppingcart/dto/customer/response/EmailDuplicateResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto.customer;
+package woowacourse.shoppingcart.dto.customer.response;
 
 public class EmailDuplicateResponse {
