+++ src/main/java/woowacourse/shoppingcart/dto/customer/response/UsernameDuplicateResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto.customer;
+package woowacourse.shoppingcart.dto.customer.response;
 
 public class UsernameDuplicateResponse {
